% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{assist}
\alias{assist}
\title{Launch the AI-powered EDA assistant}
\usage{
assist(data, chat = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{chat}{A \link[ellmer:Chat]{ellmer::Chat} instance (e.g., \code{ellmer::chat_ollama()}).
Be aware that any \code{system_prompt} will be overwritten.}
}
\description{
Launch the AI-powered EDA assistant
}
\examples{

data(diamonds, package = "ggplot2")
assist(diamonds)
}
