% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{assist}
\alias{assist}
\title{Launch the AI-powered EDA assistant}
\usage{
assist(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Launch the AI-powered EDA assistant
}
\examples{

data(diamonds, package = "ggplot2")
assist(diamonds)
}
